use float(abs(c)) to fix new numpy complaint of "same_kind" operations
Former-commit-id: ee63ed7153344885c15e054a8c44564af8b4ec0a [formerly 149b360ce0087ba89eedb44a19642a664d63699d] [formerly f05b5d90a47a1d2d1a494e089987b381a83a4ad3 [formerly fbb0971937f40f01b59de401f5db47d2b3660384]] [formerly 6fb6e030c315673e2fe6494e918f23eab303dce2 [formerly 292e4ff9512a9df20a3082bbee69d05fc196027b] [formerly 9456932224bd255e9c135d01c082bcc03f4d2807 [formerly 361373ae417efeb9037a1863c741d82b21ba6a18]]]
Former-commit-id: 0bed6f860adac5362cec193bb773f79a772d4d60 [formerly d4a8cd220c4aaff03eed06771604d7a60750139e] [formerly 6a9abb783235713e4acff8a1238a69ffa501ba18 [formerly 673342a431ab713308bb21b5da929eda002fa60a]]
Former-commit-id: cf534c75dd4a6418c7e981871e4f740aed160a45 [formerly 040a59ae267739bc07710674c79e710a2d895f44]
Former-commit-id: fc4542032a1fcf4b349ed772d66ee11cf631b9e4

+++ amset/core.py
@@ -489,7 +489,7 @@ class Amset(object):
                     self.mobility[tp]['seebeck'][c][T] *= -1e6 * k_B
                     self.mobility[tp]["seebeck"][c][T] += \
                         1e6 * self.mobility[tp]["J_th"][c][T]\
-                        /(self.mobility[tp]["overall"][c][T]*e*abs(c))/dTdz
+                        /(self.mobility[tp]["overall"][c][T]*e*float(abs(c)))/dTdz
                     for band in list(self.valleys[tp].keys()):
                         for valley_k in list(self.valleys[tp][band].keys()):
                             self.valleys[tp][band][valley_k]["seebeck"][c][T] -= (self.fermi_level[c][T] - self.cbm_vbm[tp]["energy"]) / (k_B * T)
@@ -544,7 +544,7 @@ class Amset(object):
         self.mo_labels = self.elastic_scats + self.inelastic_scats + ['overall', 'average']
         self.spb_labels = ['SPB_ACD']
         self.transport_labels = self.mo_labels + self.spb_labels + ["seebeck","J_th"]
-        self.mobility = {tp: {el_mech: {c: {T: np.array([0., 0., 0.]) \
+        self.mobility = {tp: {el_mech: {c: {T: np.array([0., 0., 0.], dtype='float') \
                                             for T in self.temperatures} \
                                         for c in self.dopings} \
                               for el_mech in self.transport_labels} \
@@ -1032,8 +1032,8 @@ class Amset(object):
         self.dos_bwidth = params.get("dos_bwidth", 0.1)
         self.nkdos = params.get("nkdos", 29)
         self.v_min = 1000
-        self.gs = 1e-32  # small value (e.g. used for an initial non-zero val)
-        self.gl = 1e32  # global large value
+        self.gs = float(1e-32)  # small value (e.g. used for an initial non-zero val)
+        self.gl = float(1e32)  # global large value
         self.BTE_iters = params.get("BTE_iters", 5)
         self.n_jobs = params.get("n_jobs", -1)
         if self.n_jobs == -1:
@@ -2471,7 +2471,7 @@ class Amset(object):
                 range_of_energies = np.arange(self.cbm_vbm[tp]['energy'] - 2,
                                               self.cbm_vbm[tp]['energy'] + 2.1,
                                               step)
-                diff = 1000 * abs(c)
+                diff = 1000.0 * abs(c)
                 while(diff > tol):
                     # try a number for fermi level
                     diffs = {}
@@ -3128,7 +3128,7 @@ class Amset(object):
                         mu_overall_valley = self.integrate_over_E(prop_list=["g"],
                                tp=tp, c=c, T=T, xDOS=False, xvel=True)
 
-                    valley_transport[tp]["J_th"][c][T] = abs(c)*e \
+                    valley_transport[tp]["J_th"][c][T] = float(abs(c))*e \
                             *self.integrate_over_E(prop_list=["g_th"],
                                                    tp=tp, c=c, T=T,
                                                    xDOS=False, xvel=True) #in A/cm2
